+++ sample/src/main/java/com/fenchtose/tooltip_demo/MainActivity.java
@@ -1,5 +1,6 @@
 package com.fenchtose.tooltip_demo;
 
+import android.content.Intent;
 import android.content.res.Resources;
 import android.support.annotation.NonNull;
 import android.support.annotation.StringRes;
@@ -13,6 +14,7 @@ import android.view.ViewGroup;
 import android.widget.TextView;
 
 import com.fenchtose.tooltip.Tooltip;
+import com.fenchtose.tooltip.TooltipAnimation;
 
 @SuppressWarnings("ConstantConditions")
 public class MainActivity extends AppCompatActivity {
@@ -25,6 +27,7 @@ public class MainActivity extends AppCompatActivity {
 
     private int tipSizeSmall;
     private int tipSizeRegular;
+    private int tipRadius;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -41,6 +44,7 @@ public class MainActivity extends AppCompatActivity {
 
         tipSizeSmall = res.getDimensionPixelSize(R.dimen.tip_dimen_small);
         tipSizeRegular = res.getDimensionPixelSize(R.dimen.tip_dimen_regular);
+        tipRadius = res.getDimensionPixelOffset(R.dimen.tip_radius);
 
         final View bottomButton = findViewById(R.id.button_bottom);
         if (bottomButton != null) {
@@ -55,7 +59,9 @@ public class MainActivity extends AppCompatActivity {
         findViewById(R.id.bottom_auto_adjust).setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                showTooltip(v, R.string.bottom_auto_adjust, Tooltip.BOTTOM, true, tooltipSize,
+                showTooltip(v, R.string.bottom_auto_adjust, Tooltip.BOTTOM, true,
+                        TooltipAnimation.SCALE,
+                        tooltipSize,
                         ViewGroup.LayoutParams.WRAP_CONTENT);
             }
         });
@@ -63,7 +69,9 @@ public class MainActivity extends AppCompatActivity {
         findViewById(R.id.bottom_no_auto_adjust).setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                showTooltip(v, R.string.bottom_no_auto_adjust, Tooltip.BOTTOM, false, tooltipSize,
+                showTooltip(v, R.string.bottom_no_auto_adjust, Tooltip.BOTTOM, false,
+                        TooltipAnimation.SCALE,
+                        tooltipSize,
                         ViewGroup.LayoutParams.WRAP_CONTENT);
             }
         });
@@ -71,7 +79,9 @@ public class MainActivity extends AppCompatActivity {
         findViewById(R.id.bottom).setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                showTooltip(v, R.string.bottom, Tooltip.BOTTOM, false, tooltipSize,
+                showTooltip(v, R.string.bottom, Tooltip.BOTTOM, false,
+                        TooltipAnimation.SCALE,
+                        tooltipSize,
                         ViewGroup.LayoutParams.WRAP_CONTENT);
             }
         });
@@ -79,7 +89,8 @@ public class MainActivity extends AppCompatActivity {
         findViewById(R.id.top_auto_adjust).setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                showTooltip(v, R.string.top_auto_adjust, Tooltip.TOP, true, tooltipSize,
+                showTooltip(v, R.string.top_auto_adjust, Tooltip.TOP, true,
+                        TooltipAnimation.NONE, tooltipSize,
                         ViewGroup.LayoutParams.WRAP_CONTENT);
             }
         });
@@ -87,7 +98,8 @@ public class MainActivity extends AppCompatActivity {
         findViewById(R.id.top_no_auto_adjust).setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                showTooltip(v, R.string.top_no_auto_adjust, Tooltip.TOP, false, tooltipSize,
+                showTooltip(v, R.string.top_no_auto_adjust, Tooltip.TOP, false,
+                        TooltipAnimation.NONE, tooltipSize,
                         ViewGroup.LayoutParams.WRAP_CONTENT);
             }
         });
@@ -95,7 +107,7 @@ public class MainActivity extends AppCompatActivity {
         findViewById(R.id.top).setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                showTooltip(v, R.string.top, Tooltip.TOP, false, tooltipSize,
+                showTooltip(v, R.string.top, Tooltip.TOP, false,TooltipAnimation.NONE, tooltipSize,
                         ViewGroup.LayoutParams.WRAP_CONTENT);
             }
         });
@@ -104,6 +116,7 @@ public class MainActivity extends AppCompatActivity {
             @Override
             public void onClick(View v) {
                 showTooltip(v, R.string.right_auto_adjust, Tooltip.RIGHT, true,
+                        TooltipAnimation.REVEAL,
                         ViewGroup.LayoutParams.WRAP_CONTENT,
                         tooltipSize);
             }
@@ -113,6 +126,7 @@ public class MainActivity extends AppCompatActivity {
             @Override
             public void onClick(View v) {
                 showTooltip(v, R.string.right_no_auto_adjust, Tooltip.RIGHT, false,
+                        TooltipAnimation.REVEAL,
                         ViewGroup.LayoutParams.WRAP_CONTENT,
                         tooltipSize);
             }
@@ -122,6 +136,7 @@ public class MainActivity extends AppCompatActivity {
             @Override
             public void onClick(View v) {
                 showTooltip(v, R.string.right, Tooltip.RIGHT, false,
+                        TooltipAnimation.REVEAL,
                         tooltipSize, ViewGroup.LayoutParams.WRAP_CONTENT);
             }
         });
@@ -130,6 +145,7 @@ public class MainActivity extends AppCompatActivity {
             @Override
             public void onClick(View v) {
                 showTooltip(v, R.string.left_auto_adjust, Tooltip.LEFT, true,
+                        TooltipAnimation.SCALE_AND_FADE,
                         ViewGroup.LayoutParams.WRAP_CONTENT,
                         tooltipSize);
             }
@@ -139,6 +155,7 @@ public class MainActivity extends AppCompatActivity {
             @Override
             public void onClick(View v) {
                 showTooltip(v, R.string.left_no_auto_adjust, Tooltip.LEFT, false,
+                        TooltipAnimation.SCALE_AND_FADE,
                         ViewGroup.LayoutParams.WRAP_CONTENT,
                         tooltipSize);
             }
@@ -148,6 +165,7 @@ public class MainActivity extends AppCompatActivity {
             @Override
             public void onClick(View v) {
                 showTooltip(v, R.string.left, Tooltip.LEFT, false,
+                        TooltipAnimation.SCALE_AND_FADE,
                         tooltipSize, ViewGroup.LayoutParams.WRAP_CONTENT);
             }
         });
@@ -161,24 +179,37 @@ public class MainActivity extends AppCompatActivity {
 
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
-        showMenuTooltip(findViewById(R.id.demo_action));
+        if (item.getItemId() == R.id.demo_action) {
+            showMenuTooltip(findViewById(R.id.demo_action));
+        } else if (item.getItemId() == R.id.second_action) {
+            openSecondActivity();
+        }
         return true;
+    }
+
+    private void openSecondActivity() {
+        Intent intent = new Intent(this, SecondActivity.class);
+        startActivity(intent);
     }
 
     private void showTooltip(@NonNull View anchor, @StringRes int resId,
                              @Tooltip.Position int position, boolean autoAdjust,
+                             @TooltipAnimation.Type int type,
                              int width, int height) {
         TextView textView = (TextView) getLayoutInflater().inflate(R.layout.tooltip_textview, null);
         textView.setText(resId);
         textView.setLayoutParams(new ViewGroup.LayoutParams(width, height));
-        showTooltip(anchor, textView, position, autoAdjust, tooltipColor);
+        showTooltip(anchor, textView, position, autoAdjust, type, tooltipColor);
     }
 
     private void showTooltip(@NonNull View anchor, @NonNull View content,
                              @Tooltip.Position int position, boolean autoAdjust,
+                             @TooltipAnimation.Type int type,
                              int tipColor) {
+
         new Tooltip.Builder(this)
                 .anchor(anchor, position)
+                .animate(new TooltipAnimation(type, 500))
                 .autoAdjust(autoAdjust)
                 .content(content)
                 .withTip(new Tooltip.Tip(tipSizeRegular, tipSizeRegular, tipColor))
@@ -190,15 +221,24 @@ public class MainActivity extends AppCompatActivity {
     private void showCustomTooltip(@NonNull View anchor) {
         View content = getLayoutInflater().inflate(R.layout.item_tooltip_view_1, null);
 
-        new Tooltip.Builder(this)
+        final Tooltip customTooltip = new Tooltip.Builder(this)
                 .anchor(anchor, Tooltip.BOTTOM)
+                .animate(new TooltipAnimation(TooltipAnimation.SCALE_AND_FADE, 400))
                 .autoAdjust(true)
                 .withPadding(tooltipPadding)
                 .content(content)
-                .withTip(new Tooltip.Tip(tipSizeRegular, tipSizeRegular, tooltipColor))
+                .cancelable(false)
+                .withTip(new Tooltip.Tip(tipSizeRegular, tipSizeRegular, tooltipColor, tipRadius))
                 .into(root)
                 .debug(true)
                 .show();
+
+        content.findViewById(R.id.dismiss_button).setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                customTooltip.dismiss(true);
+            }
+        });
     }
 
     private void showMenuTooltip(@NonNull View anchor) {
@@ -207,7 +247,9 @@ public class MainActivity extends AppCompatActivity {
 
         new Tooltip.Builder(this)
                 .anchor(anchor, Tooltip.BOTTOM)
+                .animate(new TooltipAnimation(TooltipAnimation.REVEAL, 400))
                 .autoAdjust(true)
+                .autoCancel(2000)
                 .content(textView)
                 .withPadding(getResources().getDimensionPixelOffset(R.dimen.menu_tooltip_padding))
                 .withTip(new Tooltip.Tip(tipSizeSmall, tipSizeSmall, tooltipColor))