+++ Changelog.md
@@ -5,3 +5,18 @@
 ### 0.1.2
 
  - Min sdk changed to 1
+ 
+### 0.1.3
+ 
+ - Animations added
+ - Tip corner radius added
+ - Min sdk changed to 11
+ 
+### 0.1.4
+ 
+ - Fixed issues with anchored view in the CoordinatorLayout.
+ 
+### 0.1.5
+ 
+ - Refactored code to use proper Builder pattern reducing method counts
+ - Skipping animation if view is detached